+++ samples/aapinstrumentsample/src/main/cpp/aapinstrumentsample.cpp
@@ -237,7 +237,7 @@ void sample_plugin_deactivate(AndroidAudioPlugin *plugin) {
     context->active = false;
 }
 
-int32_t sample_plugin_get_state_size(AndroidAudioPluginExtensionTarget target) {
+size_t sample_plugin_get_state_size(AndroidAudioPluginExtensionTarget target) {
     // FIXME: implement
     return 0;
 }
